requirement.txt
backports.csv==1.0.7
certifi==2019.3.9
contextlib2==0.5.5
Cython==0.29.6
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-filter==2.1.0
django-import-export==1.2.0
django-suit==2.0a1
et-xmlfile==1.0.1
get==0.0.36
jdcal==1.4
lxml==4.3.2
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
post==0.0.24
public==0.0.63
pytz==2018.9
PyYAML==5.1
query-string==0.0.26
request==0.0.24
tablib==0.13.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
